Create Assignment_09
Implement the checkInterval method. The method should check whether an integer is between 50 and 100, not inclusive, and report the result on the screen in the following form: "The number a is not in the interval." or "The number a is in the interval.", where a is the method argument.

Example for 112:
The number 112 is not in the interval.

Example for 60:
The number 60 is in the interval.